Add notifications to trayicon
- A travez del patron observer se creo una clase Notifier
- PrinterServer usa una intancia de Notifier para notificar a
  subscribers
- TrayIcon se subscribe a los eventos de PrintServer mediante Notifier
- Notifier subscribe automaticamente un logger por defecto

issue #14

+++ src/main/java/pe/puyu/pukahttp/services/api/PrinterApiController.java
@@ -15,6 +15,7 @@ import pe.puyu.pukahttp.repository.TicketRepository;
 import pe.puyu.pukahttp.services.printer.Printer;
 import pe.puyu.pukahttp.services.printer.SweetTicketPrinter;
 import pe.puyu.pukahttp.util.AppUtil;
+import pe.puyu.pukahttp.util.Notifier;
 
 import java.io.OutputStream;
 import java.text.ParseException;
@@ -38,7 +39,7 @@ public class PrinterApiController {
 		ctx.json(response);
 	}
 
-	public void openDrawer(Context ctx) {
+	public void openDrawer(Context ctx, Notifier serverNotifier) {
 		var response = new ResponseApi<String>();
 		ctx.async(
 			10000,
@@ -46,6 +47,7 @@ public class PrinterApiController {
 				response.setStatus("error");
 				response.setMessage("Error al abrir caja");
 				response.setError("Tiempo espera agotado");
+				serverNotifier.error("Error al abrir caja registradora, razon: timeout");
 				ctx.json(response);
 			},
 			() -> {
@@ -61,6 +63,7 @@ public class PrinterApiController {
 						response.setStatus("success");
 						ctx.json(response);
 					}
+					serverNotifier.info("Solicitud exitosa para abrir caja registradora.");
 				} catch (Exception e) {
 					throw new RuntimeException(e);
 				}
@@ -68,7 +71,7 @@ public class PrinterApiController {
 		);
 	}
 
-	public void printTickets(Context ctx) {
+	public void printTickets(Context ctx, Notifier serverNotifier) {
 		var response = new ResponseApi<Integer>();
 		ctx.async(
 			15000,
@@ -78,10 +81,11 @@ public class PrinterApiController {
 				response.setError("Trabajo de impresion excedio el tiempo de espera.");
 				setResponseItemsQueue(response);
 				ctx.json(response);
+				serverNotifier.error("Error al imprimir ticktes, razon: timeout");
 			},
 			() -> {
 				JsonArray tickets = JsonParser.parseString(ctx.body()).getAsJsonArray();
-				printJob(tickets);
+				printJob(tickets, serverNotifier);
 				response.setStatus("success");
 				response.setMessage("Trabajo de impresión no lanzo ningun error.");
 				setResponseItemsQueue(response);
@@ -90,26 +94,33 @@ public class PrinterApiController {
 		);
 	}
 
-	public void deleteTickets(Context ctx) throws ParseException {
+	public void deleteTickets(Context ctx, Notifier serverNotifier) throws ParseException {
 		var response = new ResponseApi<Integer>();
 		response.setData(0);
 		response.setMessage("Se libero todos tickets en memoria");
 		response.setStatus("success");
 		if (ticketRepository != null) {
+			var countTickets = ticketRepository.countAll();
 			var dateString = ctx.queryParam("date");
 			if (dateString != null && !dateString.isEmpty()) {
 				var date = AppUtil.getDateTimeFormatter().parse(dateString);
 				var rowsDeleted = ticketRepository.deleteWithDatesBefore(date);
-				response.setMessage(String.format("Se eliminaron %d tickets en memoria.", rowsDeleted));
+				if(rowsDeleted > 0){
+					response.setMessage(String.format("Se eliminaron %d tickets en memoria.", rowsDeleted));
+					serverNotifier.info(String.format("Se eliminaron %d tickets", rowsDeleted));
+				}
 			} else {
-				ticketRepository.deleteAll();
+				if (countTickets > 0) {
+					ticketRepository.deleteAll();
+					serverNotifier.info("Se eliminaron todos los tickets almacenados");
+				}
 			}
-			response.setData(ticketRepository.countAll());
+			response.setData(countTickets);
 		}
 		ctx.json(response);
 	}
 
-	public void reprintTickets(Context ctx) {
+	public void reprintTickets(Context ctx, Notifier serverNotifier) {
 		var response = new ResponseApi<Integer>();
 		ctx.async(
 			20000,
@@ -119,6 +130,7 @@ public class PrinterApiController {
 				response.setError("Trabajo de impresion excedio el tiempo de espera.");
 				setResponseItemsQueue(response);
 				ctx.json(response);
+				serverNotifier.error("Error al reimprimir tickets, razon: timeout");
 			},
 			() -> {
 				if (ticketRepository != null) {
@@ -128,7 +140,7 @@ public class PrinterApiController {
 						tickets.add(JsonParser.parseString(item.getData()));
 					}
 					ticketRepository.deleteAll();
-					printJob(tickets);
+					printJob(tickets, serverNotifier);
 					response.setData(ticketRepository.countAll());
 				} else {
 					response.setData(0);
@@ -140,7 +152,7 @@ public class PrinterApiController {
 		);
 	}
 
-	public void genericErrorHandling(Exception e, Context ctx) {
+	public void genericErrorHandling(Exception e, Context ctx, Notifier serverNotifier) {
 		logger.error("Error PrinterApiController, \nMESSAGE: {}.\nCAUSE: {}.\nLOCALIZED_MESSAGE: {}."
 			, e.getMessage()
 			, e.getCause()
@@ -151,24 +163,33 @@ public class PrinterApiController {
 		response.setStatus("error");
 		setResponseItemsQueue(response);
 		ctx.json(response);
+		serverNotifier.warn(String.format("Exception into server: %s", e.getLocalizedMessage()));
 	}
 
-	public void printJob(JsonArray tickets) {
+	public void printJob(JsonArray tickets, Notifier serverNotifier) {
 		var errors = new LinkedList<String>();
 		for (var ticket : tickets) {
+			var printer = ticket.getAsJsonObject().getAsJsonObject("printer");
 			try {
 				logger.trace("Ticket a imprimir: {}", ticket);
 				var sweetTicketPrinter = new SweetTicketPrinter(ticket.getAsJsonObject());
 				sweetTicketPrinter.setOnUncaughtException(e -> {
 					errors.add(e);
-					logger.warn("UncaughtException printJob: {}", e);
 					secureInsertTicket(ticket);
+					serverNotifier.warn(String.format("UncaughtException printJob: %s", e));
 				});
 				sweetTicketPrinter.printTicket();
 			} catch (Exception e) {
 				errors.add(e.getMessage());
 				secureInsertTicket(ticket);
 				logger.error("Error al imprimir un ticket:  {}", e.getMessage());
+				serverNotifier.error(String.format(
+					"Error al imprimir name_system: %s, port: %d, type: %s, error: %s",
+					printer.get("name_system").getAsString(),
+					printer.get("port").getAsInt(),
+					printer.get("type").getAsString(),
+					e.getLocalizedMessage()
+				));
 			}
 		}
 		if (!errors.isEmpty()) {
